# test: speed up ci test run & reduce random failures