--- a/packages/collector/test/tracing/cloud/aws-sdk/v3/dynamodb/dynamodb_test.js
+++ b/packages/collector/test/tracing/cloud/aws-sdk/v3/dynamodb/dynamodb_test.js
@@ -197,6 +197,7 @@ mochaSuiteFn('tracing/cloud/aws-sdk/v3/dynamodb', function () {
         });
 
         verifyResponse(response, availableOperations[0], false, tableName);
+        agentControls.clearReceivedData();
       });
 
       describe('attempt to get result', () => {
@@ -250,6 +251,7 @@ mochaSuiteFn('tracing/cloud/aws-sdk/v3/dynamodb', function () {
         });
 
         verifyResponse(response, availableOperations[0], false, tableName);
+        agentControls.clearReceivedData();
       });
 
       describe('attempt to get result', () => {