--- a/packages/collector/test/tracing/cloud/aws-sdk/v3/s3/client-s3_test.js
+++ b/packages/collector/test/tracing/cloud/aws-sdk/v3/s3/client-s3_test.js
@@ -65,7 +65,7 @@ mochaSuiteFn('tracing/cloud/aws-sdk/v3/s3', function () {
   describe(`version: ${version}`, function () {
     this.timeout(config.getTestTimeout() * 3);
 
-    let bucketName = 'nodejs-team';
+    let bucketName = 'nodejs-bucket';
 
     if (process.env.AWS_S3_BUCKET_NAME) {
       bucketName = `${process.env.AWS_S3_BUCKET_NAME}v3-${semver.major(process.versions.node)}-${uuid()}`;