… package.json
… packages/autoprofile/reporter-config.json
… packages/aws-fargate/package.json
… packages/aws-fargate/reporter-config.json
… packages/aws-lambda-auto-wrap/package.json
… packages/aws-lambda/package.json
… packages/aws-lambda/reporter-config.json
… packages/aws-lambda/test/integration_test/callback_test.js
… packages/aws-lambda/test/integration_test/legacy_test.js
… packages/aws-lambda/test/integration_test/promise_test.js
… packages/collector/package.json
… packages/collector/reporter-config.json
… packages/collector/test/tracing/cloud/aws-sdk/v3/dynamodb/test.js
… packages/collector/test/tracing/cloud/aws-sdk/v3/dynamodb/v3-callback/dynamodb2_test.js
… packages/collector/test/tracing/cloud/aws-sdk/v3/sqs/client-sqs2_test.js
… packages/collector/test/tracing/cloud/aws-sdk/v3/sqs/client-sqs_test.js
… packages/collector/test/tracing/cloud/aws-sdk/v3/sqs/test.js
… packages/collector/test/tracing/protocols/graphql/14_test.js
… packages/collector/test/tracing/protocols/graphql/15_test.js
… packages/collector/test/tracing/protocols/graphql/16_test.js
… packages/core/package.json
… packages/core/reporter-config.json
… packages/google-cloud-run/package.json
… packages/google-cloud-run/reporter-config.json
… packages/legacy-sensor/package.json
… packages/legacy-sensor/reporter-config.json
… packages/metrics-util/package.json
… packages/metrics-util/reporter-config.json
… packages/opentelemetry-exporter/package.json
… packages/opentelemetry-exporter/reporter-config.json