@basti1302 2022-03-31
LGTM 👍 

One thing that we should discuss IMO: There is a good bit of code duplication now to make the distribution algorithm of CircleCI happy. But I think we can eat the cake and have it too, in this case -- I added #518 as a suggestion for a pattern that should make CircleCI happy and avoid code duplication. 

Let's have a chat tomorrow or next week.